Error handling added to trade functionality.

+++ backend/src/main/java/com/techelevator/model/Balance.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public class Balance {
 
-    private long balanceId;
+    private int balanceId;
     private int gameId;
     private String username;
     private BigDecimal amount = new BigDecimal(100000);
@@ -17,7 +17,7 @@ public class Balance {
     }
 
     //Constructor for pulling information from DB
-    public Balance(long balanceId, int gameId, String username, BigDecimal amount) {
+    public Balance(int balanceId, int gameId, String username, BigDecimal amount) {
         this.balanceId = balanceId;
         this.gameId = gameId;
         this.username = username;
